Reorganize scripts: move create_disk.sh into scripts/, add wrapper; remove generated disk and build artifacts

+++ uart/uart.h
@@ -1,11 +1,13 @@
 #ifndef UART_H
 #define UART_H
 
+#include <stdint.h>
+
 // function prototypes
 void uart_init(void);
 void uart_putc(char c);                      // put a single character
 void uart_puts(const char* s);               // put a c-style string (null-terminated)
-void uart_print_hex(unsigned long value);    // print a value in hexadecimal
-void uart_print_dec(unsigned long value);    // print a value in decimal
+void uart_print_hex(uint64_t value);         // print a value in hexadecimal
+void uart_print_dec(uint32_t value);         // print a value in decimal
 
 #endif
